Game.java - Added methods for spinning the wheel that calls the player.spinWheel. It provides logic for changing the player over to the next player and bankrupting that player as well if they spin lose a turn (0) or bankrupt(-1). It also returns the value of the wheelspin to the client so it can print out messages accordingly.
added pickLetter that does input validation, guesses the letters, multiplies the occurrences by the wheel value and adds it to the player's earnings.

added buyAVowel method that does input validation as well as validation to ensure the player has enough money and buys the vowel. Still need to have an exception for it.

I figure exception handling for the validation can be done in client ie: try pickLetter catch try again or something similar